Fixed a warning

+++ src/ppu.h
@@ -40,12 +40,12 @@ struct PPU_STATUS
 
 struct PPU_SCROLL_ADDRESS
 {
-    uint8_t coarse_x : 5;
-    uint8_t coarse_y : 5;
-    uint8_t nametable_select : 2;
-    uint8_t fine_y : 3;
+    uint16_t coarse_x : 5;
+    uint16_t coarse_y : 5;
+    uint16_t nametable_select : 2;
+    uint16_t fine_y : 3;
 
-    uint8_t padding : 1;
+    uint16_t padding : 1;
 } __attribute__((packed));
 
 struct SPRITE_ATTRIBUTES
